requirements.txt
browsermob-proxy
selenium
PyYAML
psutil
colorama
termcolor